+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All notable changes to the Wazuh ML Commons project will be documented in this f
 - Added URL-based rule navigation from Log Test detection results [#308](https://github.com/wazuh/wazuh-dashboard-security-analytics/pull/308)
 - Added `opensearch_security_analytics.disabledSettings` configuration to hide specific settings in the UI about Index discarded/unclassified events [#350](https://github.com/wazuh/wazuh-dashboard-security-analytics/pull/350) [#375](https://github.com/wazuh/wazuh-dashboard-security-analytics/pull/375)
 - Added date formatter that respects the dateFormat advanced setting [#362](https://github.com/wazuh/wazuh-dashboard-security-analytics/pull/362)
+- Added a callout to promote pending changes between spaces [#381](https://github.com/wazuh/wazuh-dashboard-security-analytics/pull/381)
 
 ### Changed